Resolve merge conflicts and rename orientation->spatialorientation
