+++ README.rst
@@ -35,6 +35,7 @@ Gyroscope                          X       X
 Battery                            X       X   X       X    X
 Native file chooser                            X       X    X
 Orientation                        X
+Spatial Orientation                X
 Audio recording                    X
 Flash                              X       X
 Wifi                                           X       X    X
