+++ plyer/__init__.py
@@ -4,10 +4,11 @@ Plyer
 
 '''
 
+
 __all__ = ('accelerometer', 'audio', 'battery', 'call', 'camera', 'compass',
            'email', 'filechooser', 'gps', 'gyroscope', 'irblaster',
-           'orientation', 'notification', 'sms', 'tts', 'uniqueid', 'vibrator',
-           'wifi')
+           'orientation', 'spatialorientation' 'notification', 'sms', 'tts',
+           'uniqueid', 'vibrator', 'wifi')
 
 __version__ = '1.2.5dev'
 
@@ -50,6 +51,9 @@ irblaster = Proxy('irblaster', facades.IrBlaster)
 
 #: Orientation proxy to :class:`plyer.facades.Orientation`
 orientation = Proxy('orientation', facades.Orientation)
+
+#: SpatialOrientation proxy to :class:`plyer.facades.SpatialOrientation`
+spatialorientation = Proxy('spatialorientation', facades.SpatialOrientation)
 
 #: Notification proxy to :class:`plyer.facades.Notification`
 notification = Proxy('notification', facades.Notification)
