+++ plyer/facades/__init__.py
@@ -8,8 +8,8 @@ Interface of all the features available.
 
 __all__ = ('Accelerometer', 'Audio', 'Battery', 'Call', 'Camera', 'Compass',
            'Email', 'FileChooser', 'GPS', 'Gyroscope', 'IrBlaster',
-           'Orientation', 'Notification', 'Sms', 'TTS', 'UniqueID', 'Vibrator',
-           'Wifi', 'Flash')
+           'Orientation', 'SpatialOrientation' 'Notification', 'Sms', 'TTS',
+           'UniqueID', 'Vibrator', 'Wifi', 'Flash')
 
 from plyer.facades.accelerometer import Accelerometer
 from plyer.facades.audio import Audio
@@ -23,6 +23,7 @@ from plyer.facades.gps import GPS
 from plyer.facades.gyroscope import Gyroscope
 from plyer.facades.irblaster import IrBlaster
 from plyer.facades.orientation import Orientation
+from plyer.facades.spatialorientation import SpatialOrientation
 from plyer.facades.notification import Notification
 from plyer.facades.sms import Sms
 from plyer.facades.tts import TTS
