+++ plyer/facades/orientation.py
@@ -36,20 +36,6 @@ class Orientation(object):
     Orientation facade.
     '''
 
-    @property
-    def orientation(self):
-        '''Property that returns values of the current device orientation
-        as a (pitch, roll, azimuth) tuple.
-        Pitch has range from -180 to 180 (X angle).
-        Roll has range from -90 to 90 (Y angle).
-        Azimuth from 0 to 360 (Z angle).
-        Returns (None, None, None) if no data is currently available.
-        '''
-        return self._get_orientation() or (None, None, None)
-
-    def _get_orientation(self):
-        raise  NotImplementedError()
-
     def set_landscape(self, reverse=False):
         '''
         Rotate the app to a landscape orientation.
@@ -88,20 +74,4 @@ class Orientation(object):
         raise NotImplementedError()
 
     def _set_sensor(self, **kwargs):
-        raise NotImplementedError()
-
-    def enable_listener(self):
-        '''Enable the orientation sensor.
-        '''
-        self._enable_listener()
-
-    def _enable_listener(self, **kwargs):
-        raise NotImplementedError()
-
-    def disable_listener(self):
-        '''Disable the orientation sensor.
-        '''
-        self._disable_listener()
-
-    def _disable_listener(self, **kwargs):
         raise NotImplementedError()
